--- src/components/SolicitacaoDeKitLanche/Escola/Relatorio/style.scss
@@ -1,26 +1,0 @@
-.title-message {
-  font-weight: bold;
-  color: #686868;
-  font-size: 14px;
-}
-
-table.table-periods {
-  margin-top: 20px;
-  tr {
-    color: #686868;
-    margin-top: 20px;
-    font-size: 14px;
-    th {
-      font-weight: normal;
-      &:not(:first-child) {
-        padding-left: 40px;
-      }
-    }
-    td {
-      font-weight: bold;
-      &:not(:first-child) {
-        padding-left: 40px;
-      }
-    }
-  }
-}
