--- src/pages/Escola/InversaoDeDiaDeCardapioPage.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-import Breadcrumb from "../../components/Shareable/Breadcrumb";
-import Container from "../../components/InversaoDeDiaDeCardapio/Container";
-import Page from "../../components/Shareable/Page/Page";
-import { HOME } from "./constants";
-
-const atual = {
-  href: "/escola/inversao-de-dia-de-cardapio",
-  titulo: "Inversão de dia de Cardápio"
-};
-
-export default () => (
-  <Page titulo={atual.titulo}>
-    <Breadcrumb home={HOME} atual={atual} />
-    <Container />
-  </Page>
-);
